+++ core/vec3D.h
@@ -12,47 +12,47 @@ class vec3D {
 
 // Everything is public. Nothing needs to be hidden here.
   public:
-    double x1;
-    double x2;
-    double x3;
+    long double x1;
+    long double x2;
+    long double x3;
 
     // Default Construcor
     vec3D();
 
     // Custom Constructor:
-    vec3D(double x1, double x2, double x3);
+    vec3D(long double x1, long double x2, long double x3);
 
     // Setting fixed numbers
-    void set_values(double x1, double x2, double x3);
+    void set_values(long double x1, long double x2, long double x3);
 
     // Overloading << operator for std::ostream
     friend std::ostream& operator<<(std::ostream& os, const vec3D& vec);
 
     // Overloading * operator
     vec3D operator*(const vec3D& other);
-    friend vec3D operator*(const vec3D& vec, double scalar);
-    friend vec3D operator*(double scalar, const vec3D& vec);
+    friend vec3D operator*(const vec3D& vec, long double scalar);
+    friend vec3D operator*(long double scalar, const vec3D& vec);
 
     // Overloading / operator
     vec3D operator/(const vec3D& other);
-    friend vec3D operator/(const vec3D& vec, double scalar);
-    friend vec3D operator/(double scalar, const vec3D& vec);
+    friend vec3D operator/(const vec3D& vec, long double scalar);
+    friend vec3D operator/(long double scalar, const vec3D& vec);
 
     // Overloading + operator
     vec3D operator+(const vec3D& other);
-    friend vec3D operator+(const vec3D& vec, const double scalar);
-    friend vec3D operator+(const double scalar, const vec3D& vec);
+    friend vec3D operator+(const vec3D& vec, const long double scalar);
+    friend vec3D operator+(const long double scalar, const vec3D& vec);
 
     // Overloading - operator
     vec3D operator-(const vec3D& other);
-    friend vec3D operator-(const vec3D& vec, const double scalar);
-    friend vec3D operator-(const double scalar, const vec3D& vec);
+    friend vec3D operator-(const vec3D& vec, const long double scalar);
+    friend vec3D operator-(const long double scalar, const vec3D& vec);
 
     // Norm for length of vector
-    double norm_sq();
+    long double norm_sq();
 
     // Dot or scalar product
-    double dot(vec3D& other);
+    long double dot(vec3D& other);
 };
 
 #endif
